* Began a refactor of scripting APIs.  * Fixed syntax error in Bell's script.

+++ trunk/src/argonms/common/character/inventory/InventoryTools.java
@@ -111,7 +111,7 @@ public class InventoryTools {
 		equipCache = new HashMap<Integer, Equip>();
 	}
 
-	public static boolean canFitEntirely(Inventory inv, int itemid, short remQty, boolean breakRechargeableStack) {
+	public static boolean canFitEntirely(Inventory inv, int itemid, int remQty, boolean breakRechargeableStack) {
 		if (remQty > 0) {
 			if (!isRechargeable(itemid) || breakRechargeableStack) {
 				//TODO: getPersonalSlotMax, but this is in argonms.common. X.X
